new controller endpoint

+++ AcccountService/src/main/java/com/banking/accountservice/config/CurrentUser.java
@@ -2,10 +2,11 @@ package com.banking.accountservice.config;
 
 import lombok.Builder;
 
+
 @Builder
 public record CurrentUser(
         String subject,
-        Integer userId,
+        Long userId,
         String role,
         String firstName,
         String lastName
